move usagerights into media API and tidy deps
